requirements.txt
numpy>=1.24
pandas>=2.0
ctrader-open-api>=0.9
protobuf==3.20.1
requests>=2.28
schedule>=1.2
python-dotenv>=1.0
